fix possible null pointer exception errors

+++ src/test/java/net/sf/jabref/FileBasedTestHelper.java
@@ -33,6 +33,9 @@ public class FileBasedTestHelper {
     public static File createTempDir(String prefix, File directory) {
         try {
             File tempFile = File.createTempFile(prefix, "", directory);
+            if(tempFile == null) {
+                return null;
+            }
 
             if (!tempFile.delete()) {
                 return null;
